Fixed a bug with the container, fixed plumber in gulpfile
If the container was wrapped with a div that had no display set, it
would show no content.
The fix involves setting the container’s display and width while also
hiding the overflow from the inner margins.
I’ve also fixed an error with the gulpfile’s plumber.

I’ve tested the grid on all the latest versions of Firefox, Safari,
Chrome, Edge, and iOS and no issues so far.