Major bug fixes and general usability improvements
- Removed overflow:hidden on .container.
Using overflow hidden and negative margins to account for the use of
border:spacing seemed like a bad idea and very hacky. I think the
solution I have come up with is better - I have replaced border-spacing
with actual borders and background clip. This allowed me to remove
overflow:hidden on the container, which results in improved usability
by allowing elements to be positioned relative to a col outside the
container. Yay.

- Renamed instances of .grid to .row. Using .row just makes more sense
to me as it is essentially housing a row of data, not a grid per se.

- Added a .row-auto class and styles. This is to take into account the
fact that table-cells are designed to fill the entire width of a
container. However, sometimes you might want that and instead want just
.col-5 in a 12 column row.
To tackle this I’ve used inline-block on the .row-auto .cols. Since it
is doubtful you would use .row-auto fully loaded with cols, I think
this is fine and won’t need much inline-block hackery, however I’ll
have to do further testing to see how it actually runs.

- Replaced minifycss with cleancss

- See sass for usage examples

 - Yes I will get round to updating the readme with jekyll sometime
once i figure out how to use it properly